+++ src/prefect/__init__.py
@@ -43,7 +43,7 @@ import prefect.blocks.notifications
 import prefect.blocks.system
 import prefect.infrastructure.process
 import prefect.infrastructure.kubernetes
-import prefect.infrastructure.docker
+import prefect.infrastructure.container
 
 # Initialize the process-wide profile and registry at import time
 import prefect.context
@@ -91,7 +91,12 @@ register_renamed_module(
 register_renamed_module(
     "prefect.docker",
     "prefect.utilities.dockerutils",
-    start_date="Feb 2023",
+    start_date="Mar 2023",
+)
+register_renamed_module(
+    "prefect.infrastructure.docker",
+    "prefect.infrastructure.container",
+    start_date="Mar 2023",
 )
 inject_renamed_module_alias_finder()
