+++ src/prefect/infrastructure/__init__.py
@@ -1,5 +1,5 @@
 from prefect.infrastructure.base import Infrastructure, InfrastructureResult
-from prefect.infrastructure.docker import DockerContainer, DockerContainerResult
+from prefect.infrastructure.container import DockerContainer, DockerContainerResult
 from prefect.infrastructure.kubernetes import (
     KubernetesClusterConfig,
     KubernetesImagePullPolicy,
